+++ src/treeView/regionTreeView/RegionTreeViewProvider.ts
@@ -7,7 +7,7 @@ import { debounce } from "../../utils/debounce";
 import { RegionTreeItem } from "./RegionTreeItem";
 
 const REFRESH_TREE_DEBOUNCE_DELAY_MS = 100;
-const HIGHLIGHT_ACTIVE_REGION_DEBOUNCE_DELAY_MS = 100;
+const AUTO_HIGHLIGHT_ACTIVE_REGION_DEBOUNCE_DELAY_MS = 100;
 
 export class RegionTreeViewProvider implements vscode.TreeDataProvider<Region> {
   private _onDidChangeTreeData = new vscode.EventEmitter<undefined>();
@@ -20,23 +20,23 @@ export class RegionTreeViewProvider implements vscode.TreeDataProvider<Region> {
     REFRESH_TREE_DEBOUNCE_DELAY_MS
   );
 
-  private highlightActiveRegionTimeout: NodeJS.Timeout | undefined;
+  private autoHighlightActiveRegionTimeout: NodeJS.Timeout | undefined;
 
   constructor(private regionStore: RegionStore, subscriptions: vscode.Disposable[]) {
     this.registerListeners(subscriptions);
-    this.debouncedHighlightActiveRegion();
+    this.debouncedAutoHighlightActiveRegion();
   }
 
   private registerListeners(subscriptions: vscode.Disposable[]): void {
     vscode.window.onDidChangeActiveTextEditor(
-      this.clearHighlightActiveRegionTimeoutIfExists.bind(this),
+      this.clearAutoHighlightActiveRegionTimeoutIfExists.bind(this),
       this,
       subscriptions
     );
     vscode.workspace.onDidChangeTextDocument(this.onDocumentChange.bind(this), this, subscriptions);
     this.regionStore.onDidChangeRegions(this.debouncedRefreshTree, this, subscriptions);
     this.regionStore.onDidChangeActiveRegion(
-      this.debouncedHighlightActiveRegion.bind(this),
+      this.debouncedAutoHighlightActiveRegion.bind(this),
       undefined,
       subscriptions
     );
@@ -46,7 +46,7 @@ export class RegionTreeViewProvider implements vscode.TreeDataProvider<Region> {
     // Cancel any existing timeout to highlight the active region; we can wait for the upcoming
     // update from RegionStore to refresh the up-to-date active region.
     if (event.document === vscode.window.activeTextEditor?.document) {
-      this.clearHighlightActiveRegionTimeoutIfExists();
+      this.clearAutoHighlightActiveRegionTimeoutIfExists();
     }
   }
 
@@ -54,41 +54,48 @@ export class RegionTreeViewProvider implements vscode.TreeDataProvider<Region> {
     this._onDidChangeTreeData.fire(undefined);
   }
 
-  private debouncedHighlightActiveRegion(): void {
-    this.clearHighlightActiveRegionTimeoutIfExists();
-    this.highlightActiveRegionTimeout = setTimeout(
-      this.highlightActiveRegion.bind(this),
-      HIGHLIGHT_ACTIVE_REGION_DEBOUNCE_DELAY_MS
+  // #region Highlighting active region
+  private debouncedAutoHighlightActiveRegion(): void {
+    this.clearAutoHighlightActiveRegionTimeoutIfExists();
+    this.autoHighlightActiveRegionTimeout = setTimeout(
+      this.autoHighlightActiveRegion.bind(this),
+      AUTO_HIGHLIGHT_ACTIVE_REGION_DEBOUNCE_DELAY_MS
     );
   }
 
-  private clearHighlightActiveRegionTimeoutIfExists(): void {
-    if (this.highlightActiveRegionTimeout) {
-      clearTimeout(this.highlightActiveRegionTimeout);
-      this.highlightActiveRegionTimeout = undefined;
+  private clearAutoHighlightActiveRegionTimeoutIfExists(): void {
+    if (this.autoHighlightActiveRegionTimeout) {
+      clearTimeout(this.autoHighlightActiveRegionTimeout);
+      this.autoHighlightActiveRegionTimeout = undefined;
     }
+  }
+
+  private autoHighlightActiveRegion(): void {
+    this.clearAutoHighlightActiveRegionTimeoutIfExists();
+    const shouldAutoHighlightActiveRegion = getGlobalRegionsViewConfigValue(
+      "shouldAutoHighlightActiveRegion"
+    );
+    if (!shouldAutoHighlightActiveRegion) {
+      return;
+    }
+    if (!isCurrentActiveVersionedDocumentId(this.regionStore.versionedDocumentId)) {
+      // The active region is from an old document version. We'll auto-highlight the active region
+      // once RegionStore fires events for the new document version.
+      return;
+    }
+    this.highlightActiveRegion();
   }
 
   private highlightActiveRegion(): void {
-    this.clearHighlightActiveRegionTimeoutIfExists();
-    const shouldHighlightActiveRegion = getGlobalRegionsViewConfigValue(
-      "shouldAutoHighlightActiveRegion"
-    );
-    if (!shouldHighlightActiveRegion) {
-      return;
-    }
-    const { activeRegion, versionedDocumentId } = this.regionStore;
+    const { activeRegion } = this.regionStore;
     if (!this.treeView || !activeRegion) {
-      return;
-    }
-    if (!isCurrentActiveVersionedDocumentId(versionedDocumentId)) {
-      // The active region is from an old document version. We'll highlight the active region once
-      // RegionStore fires events for the new document version.
       return;
     }
     this.treeView.reveal(activeRegion, { select: true, focus: false });
   }
+  // #endregion
 
+  // #region Required TreeDataProvider methods
   getTreeItem(region: Region): vscode.TreeItem {
     return new RegionTreeItem(region);
   }
@@ -104,8 +111,32 @@ export class RegionTreeViewProvider implements vscode.TreeDataProvider<Region> {
   getChildren(element?: Region): Region[] {
     return element ? element.children : this.regionStore.topLevelRegions;
   }
+  // #endregion
 
+  /**
+   * Sets the tree view for this provider and registers event listeners for expand/collapse events.
+   * To be called after the tree view is created (which requires the provider to be created first,
+   * hence why this can't go in the constructor).
+   */
   setTreeView(treeView: vscode.TreeView<Region>): void {
     this.treeView = treeView;
+  }
+
+  expandAllTreeItems(): void {
+    if (!this.treeView) {
+      return;
+    }
+    for (const topLevelRegion of this.regionStore.topLevelRegions) {
+      this.treeView.reveal(topLevelRegion, {
+        select: false,
+        focus: false,
+        expand: 3, // Max depth
+      });
+    }
+    // Finish by highlighting the cursor's active region. We do this regardless of the
+    // `shouldAutoHighlightActiveRegion` setting, since the view is open anyway when/after calling
+    // Expand All, so there's no harm in revealing. This helps re-orient instead of scroll position
+    // being reset to the top of the tree view.
+    this.highlightActiveRegion();
   }
 }